+++ docs/changelog.rst
@@ -12,6 +12,7 @@ v0.2.0
 - Add :attr:`~BoilerplateInfo.variable_input_mode` option to change behaviour of variables input prompt
 - Add support for :ref:`template paths <guide-templating--template-paths>`
 - Add support for :ref:`dynamic generation <guide-dynamic-generation>`
+- Add support for installation of boilerplates from git repositories
 
 **Enhancements and Changes**
 
@@ -20,6 +21,7 @@ v0.2.0
 - :program:`prept info` now shows the basic details (name, required/optional, summary) of template variables
 - :attr:`~BoilerplateInfo.template_provider` now takes spec in standard Python module format i.e. ``module:object``
 - Output directories created by Prept are now properly cleaned up in case of errors during generation
+- ``.git`` directory is now ignored at installation time.
 
 **Fixes**